requirements.txt
huggingface_hub
Langchain
StreamLit
python-dotenv
pypdf
transformers
Langchain-huggingface
langchain-google-genai
fpdf

-e.
